+++ src/lib/src/core/db.rs
@@ -1,9 +1,2 @@
-pub mod df_db;
-pub mod index_db;
-pub mod kv_db;
-pub mod opts;
-pub mod path_db;
-pub mod str_json_db;
-pub mod str_val_db;
-pub mod tree_db;
-pub mod workspace_df_db;
+pub mod data_frames;
+pub mod key_val;
