+++ src/lib/src/core/df/sql.rs
@@ -1,6 +1,6 @@
 use polars::frame::DataFrame;
 
-use crate::{core::db::df_db, error::OxenError};
+use crate::{core::db::data_frames::df_db, error::OxenError};
 
 pub fn query_df(sql: String, conn: &mut duckdb::Connection) -> Result<DataFrame, OxenError> {
     let df = df_db::select_str(conn, sql, false, None, None)?;
