+++ src/lib/src/core/index/commit_metadata_db.rs
@@ -12,7 +12,7 @@ use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
 use crate::constants::{CACHE_DIR, HISTORY_DIR};
-use crate::core::db::df_db;
+use crate::core::db::data_frames::df_db;
 use crate::error::OxenError;
 use crate::model::metadata::to_duckdb_sql::ToDuckDBSql;
 use crate::model::{Commit, DirMetadataItem, LocalRepository};
