+++ src/lib/src/core/index/legacy_commit_entry_reader.rs
@@ -9,7 +9,7 @@ use crate::util;
 use rocksdb::{DBWithThreadMode, MultiThreaded};
 use std::path::{Path, PathBuf};
 
-use crate::core::db::path_db;
+use crate::core::db::key_val::path_db;
 use crate::model::LocalRepository;
 
 pub struct LegacyCommitEntryReader {
@@ -46,7 +46,7 @@ impl LegacyCommitEntryReader {
         commit_id: &str,
     ) -> Result<LegacyCommitEntryReader, OxenError> {
         let path = Self::db_path(&base_path, commit_id);
-        let opts = db::opts::default();
+        let opts = db::key_val::opts::default();
         log::debug!(
             "CommitEntryReader::new_from_path() commit_id: {} path: {:?}",
             commit_id,
