+++ src/lib/src/core/index/object_db_reader.rs
@@ -1,6 +1,7 @@
 use crate::constants::{self};
-use crate::core::db::tree_db::{TreeObject, TreeObjectChild};
-use crate::core::db::{self, tree_db};
+use crate::core::db;
+use crate::core::db::key_val::tree_db;
+use crate::core::db::key_val::tree_db::{TreeObject, TreeObjectChild};
 
 use crate::error::OxenError;
 
@@ -73,7 +74,7 @@ impl ObjectDBReader {
             }
         }
 
-        let opts = db::opts::default();
+        let opts = db::key_val::opts::default();
 
         Ok(Arc::new(ObjectDBReader {
             files_db: DBWithThreadMode::open_for_read_only(
