+++ src/lib/src/core/index/stager.rs
@@ -6,8 +6,9 @@
 
 use crate::api;
 use crate::constants;
-use crate::core::db::path_db;
-use crate::core::db::{self, str_json_db};
+use crate::core::db;
+use crate::core::db::key_val::path_db;
+use crate::core::db::key_val::str_json_db;
 use crate::core::df::tabular;
 use crate::core::index::oxenignore;
 use crate::core::index::ObjectDBReader;
@@ -91,7 +92,7 @@ impl Stager {
         let dir_db_path = Stager::dirs_db_path(&repository.path)?;
         let schemas_db_path = Stager::schemas_db_path(&repository.path)?;
 
-        let opts = db::opts::default();
+        let opts = db::key_val::opts::default();
         Ok(Stager {
             dir_db: DBWithThreadMode::open(&opts, dunce::simplified(&dir_db_path))?,
             schemas_db: DBWithThreadMode::open(&opts, dunce::simplified(&schemas_db_path))?,
@@ -104,7 +105,7 @@ impl Stager {
         let dir_db_path = Stager::dirs_db_path(&repository.path)?;
         let schemas_db_path = Stager::schemas_db_path(&repository.path)?;
 
-        let opts = db::opts::default();
+        let opts = db::key_val::opts::default();
         Ok(Stager {
             dir_db: DBWithThreadMode::open(&opts, dunce::simplified(&dir_db_path))?,
             schemas_db: DBWithThreadMode::open(&opts, dunce::simplified(&schemas_db_path))?,
