+++ src/lib/src/core/index/workspaces/stager.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 
 use crate::constants::{FILES_DIR, MODS_DIR};
 use crate::core::db;
-use crate::core::db::str_json_db;
+use crate::core::db::key_val::str_json_db;
 use crate::core::index::{CommitEntryReader, Stager};
 use crate::error::OxenError;
 use crate::model::workspace::Workspace;
@@ -56,7 +56,7 @@ fn list_staged_entries(workspace: &Workspace, status: &mut StagedData) -> Result
 pub fn list_files(workspace: &Workspace) -> Result<Vec<PathBuf>, OxenError> {
     let db_path = files_db_path(workspace);
     log::debug!("list_entries from files_db_path {db_path:?}");
-    let opts = db::opts::default();
+    let opts = db::key_val::opts::default();
     let db: DBWithThreadMode<SingleThreaded> = rocksdb::DBWithThreadMode::open(&opts, db_path)?;
     str_json_db::list_vals(&db)
 }
@@ -65,14 +65,14 @@ pub fn add(workspace: &Workspace, path: impl AsRef<Path>) -> Result<(), OxenErro
     let path = path.as_ref();
     let db_path = files_db_path(workspace);
     log::debug!("workspaces::stager::add to db_path {db_path:?}");
-    let opts = db::opts::default();
+    let opts = db::key_val::opts::default();
     let db: DBWithThreadMode<MultiThreaded> = rocksdb::DBWithThreadMode::open(&opts, db_path)?;
     let key = path.to_string_lossy();
     str_json_db::put(&db, &key, &key)
 }
 
 pub fn rm(workspace: &Workspace, path: impl AsRef<Path>) -> Result<(), OxenError> {
-    let opts = db::opts::default();
+    let opts = db::key_val::opts::default();
     let files_db_path = files_db_path(workspace);
     let files_db: DBWithThreadMode<MultiThreaded> =
         rocksdb::DBWithThreadMode::open(&opts, files_db_path)?;
