+++ src/server/src/services/workspaces/data_frames/columns.rs
@@ -13,16 +13,16 @@ pub fn rows() -> Scope {
             "/resource/{path:.*}",
             web::post().to(controllers::workspaces::data_frames::columns::create),
         )
-        // .route(
-        //     "/{row_id}/resource/{path:.*}",
-        //     web::put().to(controllers::workspaces::data_frames::rows::update),
-        // )
-        // .route(
-        //     "/{row_id}/resource/{path:.*}",
-        //     web::delete().to(controllers::workspaces::data_frames::rows::delete),
-        // )
-        // .route(
-        //     "/{row_id}/resource/{path:.*}",
-        //     web::get().to(controllers::workspaces::data_frames::rows::get),
-        // )
+    // .route(
+    //     "/{row_id}/resource/{path:.*}",
+    //     web::put().to(controllers::workspaces::data_frames::rows::update),
+    // )
+    // .route(
+    //     "/{row_id}/resource/{path:.*}",
+    //     web::delete().to(controllers::workspaces::data_frames::rows::delete),
+    // )
+    // .route(
+    //     "/{row_id}/resource/{path:.*}",
+    //     web::get().to(controllers::workspaces::data_frames::rows::get),
+    // )
 }
